CWNT_: add single frame full preview

+++ src-plugins/CWNT_/fiji/plugin/cwnt/segmentation/CWNTLivePreviewer.java
@@ -134,7 +134,7 @@ public class CWNTLivePreviewer extends MouseAdapter implements ActionListener {
 		double snipPixels = snip.getWidth() * snip.getHeight() * snip.getNSlices() * snip.getNFrames();
 		double allPixels = imp.getWidth() * imp.getHeight() * imp.getNSlices() * imp.getNFrames();
 		double dt = nucleiMasker.getProcessingTime();
-		double tmin = Math.ceil(dt * allPixels / snipPixels / 1e3 / 60); //min 
+		int tmin = (int) Math.ceil(dt * allPixels / snipPixels / 1e3 / 60); //min 
 		source.labelDurationEstimate.setText("Total duration rough estimate: "+tmin+" min.");
 
 		// Prepare results holder;
